% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtcassignment.R
\name{rtc_to_results}
\alias{rtc_to_results}
\title{An R Package for Assigning Results from MALDI-TOF Real-Time Classification (RTC) of Bacterial Isolates}
\usage{
rtc_to_results(rtc_input_name, result_name)
}
\arguments{
\item{rtc_input_name}{Path & name your input file.}

\item{result_name}{Path & name for your desired output file}
}
\value{
Writes a text file as output with the assignemnt results
}
\description{
This is a function for RTC result assignment. MALDI-TOF results for microbial isolates are generated in a PDF format, with embedded table of results.
Four lines of results for each isolate is provided.
This package takes the result table in a text file format, applies predetermined logic for processing those results and assignes classes, and tabulates the assignment.

What this program does:

Step1: Accepts the raw data and cleans it up.

Step2: Creates a flat file (each id on one line). This is the intermed file object.

Step3: Uses part of the intermediate file as final input, to generate the result file, based on logic provided.

Step4: Writes a text file as output with the results

Input: A text file (tax separated) with 3 columns, id, organism, and score.

Note: Orgnasim id should be in this format: Genus species_strain

To run this function from command line, ensure you have the source code file in your working directory & run the following to load:
source("rtcassignment.R")


Computational Requirements
This program is written for easy implementation, not speed.
It uses base R code, no libraries required, so can be run on any machine with R installed.
}
\examples{
rtc_to_results("inst/extdata/example_data.txt", "example_assignment.txt")
}
